requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.4.0
httpx==0.26.0

# AI/LLM
anthropic==0.18.1

# Document Handling
python-docx==1.1.0
pypdf2==3.0.1
python-multipart==0.0.6

# Caching & Background Jobs
redis==5.0.1
celery==5.3.6
flower==2.0.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Type Checking
mypy==1.8.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
